Version 0.33+exmpl
v0.33 with an example workspace.
On unzipping the zip file to install there will be an additional zip file 'wksExample.zip'. Move this somewhere else and unzip it. Then  cd to the new unzipped directory 'wksExample'.
Enter vim ( or gvim) then run the command ':WKSpace example.vimwks'

The example is only for the gnu toolset, but could easily be modified for other toolsets.